modoo_monitoring_dashboard/requirements.txt
altair==5.5.0
attrs==25.3.0
bcrypt==4.3.0
blinker==1.9.0
cachetools==6.1.0
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.3
click==8.2.1
cryptography==45.0.6
gitdb==4.0.12
GitPython==3.1.45
idna==3.10
invoke==2.2.0
Jinja2==3.1.6
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
MarkupSafe==3.0.2
narwhals==2.1.1
numpy==2.3.2
packaging==25.0
pandas==2.3.1
paramiko==4.0.0
pillow==11.3.0
protobuf==6.31.1
pyarrow==21.0.0
pycparser==2.22
pydeck==0.9.1
PyMySQL==1.1.1
PyNaCl==1.5.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
referencing==0.36.2
requests==2.32.4
rpds-py==0.27.0
six==1.17.0
smmap==5.0.2
SQLAlchemy==2.0.43
sshtunnel==0.4.0
streamlit==1.48.0
tenacity==9.1.2
toml==0.10.2
tornado==6.5.2
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
